aux_gen: Add remarks config to [[rule]] (#495)
## Description

Adds a new, optional, field to the [[rule]] portion of the configuration
file for generating an aux. This field allows users to add additional
remarks / comments / context around the rule. This metadata is
transfered to the raw data json coverage report.

```toml
[[rule]]
remarks = "I am adding a remark"
```

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Validated the remarks is transfered to the raw data json coverage
report.

## Integration Instructions

N/A