[CHERRY-PICK] Remove PCI related content from this repository
This change is meant to remove the direct reliance on PCI related
content because the code in this repository does not use it directly,
thus not introducing unneeded dependencies.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

This is not a functional change.

N/A
(cherry picked from commit d5bc4a51babec836c78aacbefc35835ad4e36fd2)