[CHERRY-PICK] Remove PCI related content from this repository (#502)
## Description

Cherry-pick of https://github.com/microsoft/mu_feature_mm_supv/pull/500

This change is meant to remove the direct reliance on PCI related
content because the code in this repository does not use it directly,
thus not introducing unneeded dependencies.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

This is not a functional change.

N/A
(cherry picked from commit d5bc4a51babec836c78aacbefc35835ad4e36fd2)

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This is not a functional change.

## Integration Instructions

N/A

Co-authored-by: kuqin12 <Kun.Qin@microsoft.com>