GenSeaArtifacts.py: Make functions more customizable
This commit changes the input parameters of the different helper
functions in GenSeaArtifacts.py to require more specific input paths,
instead of assuming build directory structure. This allows for more
flexibility in how the functions can be used, in scenarios where users
wish to use these functions outside of the standard build process, or
when multiple simultaneous builds of the same component cause the build
directory structure to differ than normal.

Additionally, this commit removes the generate_sea_artifacts function
as much of its functionality was outside the scope of this tool
(specifically the copying of artifacts to specific directories outside
of the caller's control). The useful parts of this functionality were
moved to standalone helper functions (generate_test_aux_binary and
generate_stm_binary) that can be used independently.