Transform PiSmmCpuDxeSmm OverrideValidation from Override to Track.
When CI runs, it uses the top of tree mu_basecore.
The override validation will continuously fail because of this.

Change override validation of PiSmmCpuDxeSmm to track tags.
Add a track tag for the latest mu_basecore, and leave track tags for the previous
version, which is still a valid target for mm supv.

When a breaking change is added to mm supv, the track tags can be deleted
and a new minimum version of track tags can be created.