Remove unused crypto functions from MbedTls libraries (#514)
## Description

We really only need the SHA256 functionality from MbedTls. Even the
linker will remove the unused functions eventually, there is still no
need to compile them if they are knowingly not getting used.

- [x] Impacts functionality?
- [x] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This was tested by passing the pipeline build that has failed due to
unresolved symbols

## Integration Instructions

N/A

---------

Co-authored-by: Aaron Pop <aaronpop@microsoft.com>