Do not include zeroed data in aux file
When generating the auxiliary file, we create a default value for each
validation entry. Most of the time, this default value is all zeros. In
a test case, over 7K worth of default values were all zeros.

This commit updates the aux file generation to not include default
values that are all zeros. Instead, we set the OffsetToDefault field
to MAX_UINT32. At runtime, when the reversion to default is done, if
the OffsetToDefault field is MAX_UINT32, we zero the memory instead of
copying from the aux file.