SeaPkg/test_aux: Add offset and size columns to results table
When given a failure like the following:

"PeCoffImageValidationMemAttr: Current entry 0x7BE13270 has invalid "
"size of 0x68, max is 0x8"

It is convenient to have the offset and size printed in the symbol
result list. You can quickly verify the offset against other
sources and confirm the size matches what is reported in the error.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>