Remove Ring 0 instructions (#542)
## Description

This is a change that removes knowing faulting instructions when the
running code is in Ring 3.

The intention is to make sure the build time failure will prevent the
#GP during runtime.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [x] Impacts functionality?
- [x] Impacts security?
- [x] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This is being tested on QEMU Q35 platform.

## Integration Instructions

N/A. If unresolved symbols are found, one needs to reach out to us to
see what it takes to update the syscall interface.