MmSupervisorPkg: Increase Mem Snap Shot static size.
In more complex platforms, there can be more page table
entries that were unblocked for Supervised MM. The static
value used when generating a snapshot, (1k), my not be
large enough to encompass the page table.

Increase the static size of the snapshot to 4k to allow
reporting of these more complex systems.