Signal `gEfiMmEndOfPeiProtocol` into MM environment (#578)
## Description

The current logic neither properly translate `gEfiEndOfPeiSignalPpiGuid`
into `gEfiMmEndOfPeiProtocol`, nor register the corresponding MMI
handler to signal the event to the rest of the user modules.

This change first added the translation logic into `MmIplPei` module,
then a MMI handler is registered through Ring3 broker which will install
a protocol as a signal to the entire user space.

The translation is done in MmIplPei specifically due to this is also
done in `StandaloneMmIplPei`.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This change is tested on QEMU Q35 platform. During the test,
`MmEndOfPeiHandler` print was observed and system booted to UEFI shell.

## Integration Instructions

N/A