Fixing failing unit test for not populating correct `MessageLength` (#580)
## Description

This change fixes a bug in the unit test where the `MessageLength` for
the corresponding MMI request is not being properly populated, causing
the test case to fail on the pipeline.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

The test passed on QEMU Q35 platform.

## Integration Instructions

N/A