SeaPkg: GenSeaArtifacts: Update scroll versions (#594)
## Description

This change updates the scroll versions inside the Cargo.toml files for
both test_aux and gen_manifest.

This is to be consistent with the scroll version specified in other
modules.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This was tested on the platform and fixed the build break.

## Integration Instructions

N/A

Signed-off-by: Kun Qin <kuqin12@gmail.com>