MmSupervisorPkg: DxeMmSupervisorVersionPublicationLib: Fix buffer size (#592)
## Description

The current implementation did not populate the message length properly.
With the hardened MMI entrypoint, this is caught by the supervisor core
and caused some asserts.

This change fixed the size calculation for supervisor version querying.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This was tested on physical hardware and no longer break the boot
process.

## Integration Instructions

N/A

Signed-off-by: Kun Qin <kuqin12@gmail.com>