MmSupervisorPkg/SeaPkg: Remove gMmSupervisorCommunicationRegionTableGuid (#591)
## Description

With the removal of MM core private data, the current module no longer
produces `gMmSupervisorCommunicationRegionTableGuid` configuration
table.

This change removed the dependencies on such table from the test
applications and instead grab it through the supervisor protocol.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This was tested with QEMU Q35.

## Integration Instructions

N/A

Signed-off-by: Kun Qin <kuqin12@gmail.com>