Skip AP validation if the registers matches BSP (#588)
## Description

This change intends to resolve a performance issue where the current
flow runs validation on all cores.

The change moves the validation to BSP routine only and APs will go over
the register validation only and verify if the supervisor base addresses
are consistent across all cores.

- [x] Impacts functionality?
- [x] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This was tested on QEMU Q35 shell based application.

## Integration Instructions

N/A

---------

Signed-off-by: kuqin12 <42554914+kuqin12@users.noreply.github.com>
Co-authored-by: Michael Kubacki <michael.kubacki@microsoft.com>